pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nestscout"
version = "0.1.0"
description = "AI-powered real estate intelligence platform with personalised property scoring"
readme = "README.md"
license = {text = "AGPL-3.0"}
requires-python = ">=3.11"
authors = [
    {name = "Jason Jiménez Cruz"},
]
dependencies = [
    # Flask & Web
    "Flask>=3.0",
    "flask-sqlalchemy>=3.1",
    "flask-migrate>=4.0",
    "flask-jwt-extended>=4.6",
    "flask-marshmallow>=1.2",
    "marshmallow-sqlalchemy>=1.0",
    "flask-cors>=4.0",
    # Database
    "SQLAlchemy>=2.0",
    "alembic>=1.13",
    "psycopg2-binary>=2.9",
    # CLI
    "click>=8.1",
    "rich>=13.0",
    "tabulate>=0.9",
    # AI / ML
    "langchain>=0.3",
    "langchain-openai>=0.3",
    "langchain-community>=0.3",
    # Data Processing
    "pandas>=2.1",
    "openpyxl>=3.1",
    # Utilities
    "python-dotenv>=1.0",
    "bcrypt>=4.1",
    "marshmallow>=3.20",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "httpx>=0.27",
    "factory-boy>=3.3",
    "ruff>=0.4",
]
ml = [
    "scikit-learn>=1.4",
    "xgboost>=2.0",
]
vectors = [
    "qdrant-client>=1.7",
    "sentence-transformers>=2.3",
]

[project.scripts]
nestscout = "nestscout.cli:cli"

[tool.setuptools.packages.find]
include = ["nestscout*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
